check fields and table existence on migration

+++ src/database/migrations/2016_12_18_115100_drop_slack_oauth.php
@@ -14,7 +14,9 @@ class DropSlackOauth extends Migration
      */
     public function up()
     {
-        Schema::drop('slack_users');
+        if (Schema::hasTable('slack_users')) {
+            Schema::drop('slack_users');
+        }
     }
 
     public function down()
