+++ src/database/migrations/2016_12_18_115100_drop_slack_users_invited.php
@@ -15,7 +15,9 @@ class DropSlackUsersInvited extends Migration
     public function up()
     {
         Schema::table('slack_users', function (Blueprint $table) {
-            $table->dropColumn('invited');
+            if (Schema::hasColumn('slack_users', 'invited')) {
+                $table->dropColumn('invited');
+            }
         });
     }
